Share CounterAddOp between tests
